the-librarians-nook/requirements.txt
gunicorn
flask
beautifulsoup4
selenium
webdriver_manager
firebase-admin
chromedriver-binary-auto
git+https://github.com/samuellando/anyREST
